requirements/tests.txt
flake8>=2.4.1
coverage==3.7.1
coveralls==0.3
django-tinymce>=1.5.3
